Fix the 54 Postgres failures, and the three product bugs behind them
Most were not test noise. Pointing the suite at the engine the fleet runs
surfaced real defects that SQLite could not express:

* PostgresStore was missing sixteen store helpers that lived on SQLiteStore
  alone -- humans, pipeline cursors, task-flow analytics, admission episodes.
  The Store protocol declared only the seven primitives, so
  isinstance(store, Store) passed while `GET /humans` returned 500 in
  production. They are now defined once in store_helpers.StoreHelpersMixin,
  which both backends inherit, and declared on the protocol so a missing one
  is a type error rather than an AttributeError at the first request.

* idx_agents_live_name, the partial index that keeps /agents from scanning
  every tombstone, existed only in the SQLite migration path. Postgres never
  had it, so the ~1s /agents latency the index was added to fix was still
  being paid in production. Ported, and the plan is now an index scan with no
  sort.

* scientific_optimizer compared a JSON boolean to the integer 1. json_extract
  yields 1/0 on SQLite but 'true'/'false' on Postgres, so the query failed
  outright with "COALESCE types text and integer cannot be matched" and the
  optimizer could sample nothing. Compared as text now, accepting both
  spellings; verified the SQLite semantics are unchanged.

Dialect fixes in service code: `col IS ?` is a SQLite null-safe comparison
that Postgres rejects, split into IS NULL / = ? branches as
provisioning_service and workflow_service already do; `INSERT OR IGNORE` in
the shared helpers became ON CONFLICT DO NOTHING.

The append-only triggers reported different text per engine -- SQLite says
"immutable" for UPDATE and "append-only" for DELETE, the port collapsed both
into "append-only" for ten tables. Aligned on TG_OP so the same operation
reports the same thing wherever it runs. (Three ref-retirement tables keep a
deliberately combined message.)

Test-side, where the intent was engine-neutral all along: assertions on
sqlite3.IntegrityError now accept StoreError too, per store.py's own guidance;
index and table introspection branches on the backend instead of assuming
PRAGMA and sqlite_master; the write-free dispatch assertion records statements
at the store API rather than through sqlite3's trace callback; and
foreign_keys_suspended() replaces a raw PRAGMA. That helper yields the
connection rather than suspending globally, because the Postgres setting is
per-session and every store.execute() borrows a different pooled connection.

CI now provisions Postgres for every job that runs tests, using the same
script developers use. A service container cannot work here:
max_locks_per_transaction is a postmaster setting and the default of 64 is
exhausted by parallel per-schema DDL.

10110 passed, 0 failed, in 590s.
