+++ docs/env-config-reference.md
@@ -1080,6 +1080,12 @@ Defaults shown as `consumer-defined` are intentionally owned by the calling subs
 | `MAC_TEST_COVERAGE` | str | consumer-defined | core | Core setting: test coverage. |
 | `MAC_TEST_DISABLE_GROUPS` | str | consumer-defined | core | Core setting: test disable groups. |
 | `MAC_TEST_JOBS` | str | 2 | core | Core setting: test jobs. |
+| `MAC_TEST_PG_CONTAINER` | str | consumer-defined | core | Core setting: test pg container. |
+| `MAC_TEST_PG_DB` | str | consumer-defined | core | Core setting: test pg db. |
+| `MAC_TEST_PG_IMAGE` | str | consumer-defined | core | Core setting: test pg image. |
+| `MAC_TEST_PG_MAX_LOCKS` | str | consumer-defined | core | Core setting: test pg max locks. |
+| `MAC_TEST_PG_PORT` | int | consumer-defined | core | Core setting: test pg port. |
+| `MAC_TEST_PG_URL` | str | consumer-defined | core | Core setting: test pg url. |
 | `MAC_TEST_PORTFOLIO` | str | consumer-defined | core | Core setting: test portfolio. |
 | `MAC_TEST_PORTFOLIO_OUTPUT` | str | consumer-defined | core | Core setting: test portfolio output. |
 | `MAC_TEST_REBUILD_MAP` | bool | consumer-defined | core | Core setting: test rebuild map. |
