+++ tests/test_source_release_store.py
@@ -29,7 +29,7 @@ from mac.models import (
     new_id,
     utcnow,
 )
-from mac.store import SQLiteStore
+from mac.store import SQLiteStore, StoreError
 
 
 # ---------------------------------------------------------------------------
@@ -437,7 +437,7 @@ class TestSQLiteConstraints:
 
     def test_duplicate_repo_sha_rejected(self, db: SQLiteStore) -> None:
         self._insert_release(db, GOOD_SHA)
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             # Same repository_id + commit_sha
             rel2 = _good_release()  # same defaults including repository_id
             _store_release(db, rel2)
@@ -445,14 +445,14 @@ class TestSQLiteConstraints:
     def test_sha_immutability_trigger(self, db: SQLiteStore) -> None:
         rel_id = self._insert_release(db)
         new_sha = "b" * 40
-        with pytest.raises(sqlite3.IntegrityError, match="immutable"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="immutable"):
             db.execute(
                 "UPDATE source_releases SET commit_sha = ? WHERE id = ?",
                 (new_sha, rel_id),
             )
 
     def test_branch_ref_check_constraint(self, db: SQLiteStore) -> None:
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             db.execute(
                 """
                 INSERT INTO source_releases
@@ -474,7 +474,7 @@ class TestSQLiteConstraints:
         rel_id = self._insert_release(db)
         dss = _good_desired_state(release_id=rel_id)
         _store_desired(db, dss)
-        with pytest.raises(sqlite3.IntegrityError, match="monotonically"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="monotonically"):
             db.execute(
                 "UPDATE fleet_desired_source_states SET generation = 1 WHERE id = ?",
                 (dss.id,),
@@ -503,7 +503,7 @@ class TestSQLiteConstraints:
             fleet_id="fleet_x",
             id=new_id("dss"),
         )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             _store_desired(db, dss2)
 
     def test_idempotency_unique_constraint(self, db: SQLiteStore) -> None:
@@ -518,7 +518,7 @@ class TestSQLiteConstraints:
             """,
             (new_id("idem"), "fleet:fleet_abc", "req_001", dss.id, 1, utcnow()),
         )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             db.execute(
                 """
                 INSERT INTO fleet_desired_source_idempotency
