+++ tests/test_work_package_schema.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import pytest
 
 from mac.models import ValidationError, WorkPackageEpoch
-from mac.store import SQLiteStore
+from mac.store import SQLiteStore, StoreError
 from mac.work_package_store import (
     get_work_package_task_link,
     guard_generic_task_mutation,
@@ -416,7 +416,7 @@ def test_ref_retirement_intent_is_append_only_and_failed_attempt_is_retryable() 
         assert store.query_one(
             "SELECT COUNT(*) AS n FROM work_package_ref_retirement_receipts"
         )["n"] == 0
-        with pytest.raises(sqlite3.IntegrityError, match="immutable"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="immutable"):
             store.execute(
                 "UPDATE work_package_ref_retirement_intents SET expected_sha = ? "
                 "WHERE id = ?",
@@ -427,7 +427,7 @@ def test_ref_retirement_intent_is_append_only_and_failed_attempt_is_retryable() 
             "id, intent_id, outcome, completed_at) VALUES (?, ?, ?, ?)",
             ("receipt_1", "intent_1", "missing", "later"),
         )
-        with pytest.raises(sqlite3.IntegrityError, match="append-only"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="append-only"):
             store.execute(
                 "DELETE FROM work_package_ref_retirement_receipts WHERE id = ?",
                 ("receipt_1",),
@@ -440,14 +440,14 @@ def test_ref_retirement_intent_is_append_only_and_failed_attempt_is_retryable() 
 def test_package_state_and_single_active_epoch_are_database_invariants() -> None:
     store = SQLiteStore(":memory:")
     try:
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             store.execute(
                 "INSERT INTO work_packages ("
                 "id, goal, state, metadata, created_by, created_at, updated_at"
                 ") VALUES (?, ?, ?, ?, ?, ?, ?)",
                 ("wp_bad", "bad", "running-ish", "{}", "human", "now", "now"),
             )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             store.execute(
                 "INSERT INTO work_packages ("
                 "id, goal, state, metadata, created_by, created_at, updated_at"
@@ -465,7 +465,7 @@ def test_package_state_and_single_active_epoch_are_database_invariants() -> None
             )
         )
         assert epoch_model.planning_base_sha == "a" * 40
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             store.execute(
                 "INSERT INTO work_package_epochs ("
                 "package_id, epoch, plan_version, planning_base_ref, planning_base_sha, status, reason, "
@@ -483,20 +483,20 @@ def test_package_state_and_single_active_epoch_are_database_invariants() -> None
                     "now",
                 ),
             )
-        with pytest.raises(sqlite3.IntegrityError, match="deactivate"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="deactivate"):
             store.execute(
                 "UPDATE work_package_epochs SET status = ? "
                 "WHERE package_id = ? AND epoch = ?",
                 ("superseded", "wp_1", 1),
             )
-        with pytest.raises(sqlite3.IntegrityError, match="state transition"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="state transition"):
             store.execute(
                 "UPDATE work_packages SET state = ? WHERE id = ?", ("draft", "wp_1")
             )
         store.execute(
             "UPDATE work_packages SET state = ? WHERE id = ?", ("completed", "wp_1")
         )
-        with pytest.raises(sqlite3.IntegrityError, match="state transition"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="state transition"):
             store.execute(
                 "UPDATE work_packages SET state = ? WHERE id = ?", ("active", "wp_1")
             )
@@ -515,7 +515,7 @@ def test_plan_versions_form_an_immutable_parent_chain() -> None:
             ") VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
             ("wp_1", 2, 1, "{}", "sha256:two", "replan", "human", "now"),
         )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             store.execute(
                 "INSERT INTO work_package_plan_versions ("
                 "package_id, version, parent_version, definition, plan_digest, reason, "
@@ -523,7 +523,7 @@ def test_plan_versions_form_an_immutable_parent_chain() -> None:
                 ") VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
                 ("wp_1", 3, 2, "{}", "sha256:one", "duplicate", "human", "now"),
             )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             store.execute(
                 "UPDATE work_package_plan_versions SET parent_version = 2 "
                 "WHERE package_id = ? AND version = 1",
@@ -575,7 +575,7 @@ def test_epoch_swap_is_atomic_and_failed_swap_rolls_back() -> None:
             )
         } == {1: "superseded", 2: "active"}
 
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             swap_work_package_epoch(
                 store,
                 package_id="wp_1",
@@ -637,14 +637,14 @@ def test_history_epoch_and_plan_version_must_identify_the_same_snapshot() -> Non
                 "now",
             ),
         )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             store.execute(
                 "INSERT INTO work_package_history ("
                 "id, package_id, seq, event_type, actor, plan_version, epoch, detail, "
                 "created_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)",
                 ("history_bad", "wp_1", 1, "bad", "human", 2, 1, "{}", "now"),
             )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             store.execute(
                 "INSERT INTO work_package_history ("
                 "id, package_id, seq, event_type, actor, plan_version, epoch, detail, "
@@ -733,20 +733,20 @@ def test_certification_is_bound_to_exact_batch_candidate() -> None:
             "agent_test",
             "now",
         )
-        with pytest.raises(sqlite3.IntegrityError, match="verifying batch"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="verifying batch"):
             store.execute(cert_sql, cert_params)
         store.execute(
             "UPDATE work_package_integration_batches SET state = ? WHERE id = ?",
             ("verifying", "batch_1"),
         )
         store.execute(cert_sql, cert_params)
-        with pytest.raises(sqlite3.IntegrityError, match="identity"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="identity"):
             store.execute(
                 "UPDATE work_package_certifications SET verification_digest = ? "
                 "WHERE id = ?",
                 ("sha256:forged", "cert_1"),
             )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             store.execute(
                 "INSERT INTO work_package_certifications ("
                 "id, batch_id, package_id, plan_version, epoch, candidate_sha, "
@@ -791,18 +791,18 @@ def test_generic_mutation_guard_and_task_link_identity_are_fail_closed() -> None
         with pytest.raises(ValidationError, match="package-aware transaction"):
             guard_generic_task_mutation(store, "task_node", "update_task")
         guard_generic_task_mutation(store, "ordinary_task", "update_task")
-        with pytest.raises(sqlite3.IntegrityError, match="identity is immutable"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="identity is immutable"):
             store.execute(
                 "UPDATE work_package_task_links SET input_digest = ? WHERE task_id = ?",
                 ("sha256:changed", "task_node"),
             )
-        with pytest.raises(sqlite3.IntegrityError, match="immutable"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="immutable"):
             store.execute(
                 "UPDATE work_package_assignment_audit SET attempt_ref = ? "
                 "WHERE lease_id = ?",
                 ("refs/mac/forged", "lease_1"),
             )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             _insert_link_and_assignment(
                 store,
                 task_id="task_bad_digest",
@@ -855,7 +855,7 @@ def test_wip_token_survives_execution_lease_expiry() -> None:
             )["state"]
             == "held"
         )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             store.execute(
                 "INSERT INTO work_package_wip_tokens ("
                 "id, package_id, plan_version, epoch, node_key, task_id, resource_key, "
@@ -884,12 +884,12 @@ def test_wip_token_survives_execution_lease_expiry() -> None:
             "release_reason = ? WHERE id = ?",
             ("released", "now", "candidate buffered", "wip_1"),
         )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             store.execute(
                 "UPDATE work_package_wip_tokens SET state = ? WHERE id = ?",
                 ("held", "wip_1"),
             )
-        with pytest.raises(sqlite3.IntegrityError, match="metadata"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="metadata"):
             store.execute(
                 "UPDATE work_package_wip_tokens SET release_reason = ? WHERE id = ?",
                 ("rewritten", "wip_1"),
@@ -998,12 +998,12 @@ def test_expired_package_task_requeue_requires_receipt_and_retains_exact_wip() -
     store = SQLiteStore(":memory:")
     try:
         decision = _prepare_expired_package_execution(store, target_state="open")
-        with pytest.raises(sqlite3.IntegrityError, match="repair receipt"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="repair receipt"):
             store.execute(
                 "UPDATE tasks SET state = 'open', owner_agent_id = NULL, "
                 "lease_id = NULL, leased_until = NULL WHERE id = 'task_node'"
             )
-        with pytest.raises(sqlite3.IntegrityError, match="exact lease-expiry repair"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="exact lease-expiry repair"):
             store.execute(
                 "UPDATE work_package_task_links SET node_state = 'ready' "
                 "WHERE task_id = 'task_node'"
@@ -1014,7 +1014,7 @@ def test_expired_package_task_requeue_requires_receipt_and_retains_exact_wip() -
             decision=decision,
             target_task_state="open",
         )
-        with pytest.raises(sqlite3.IntegrityError, match="exact lease-expiry repair"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="exact lease-expiry repair"):
             store.execute(
                 "UPDATE work_package_task_links SET node_state = 'ready' "
                 "WHERE task_id = 'task_node'"
@@ -1030,7 +1030,7 @@ def test_expired_package_task_requeue_requires_receipt_and_retains_exact_wip() -
         assert store.query_one(
             "SELECT state FROM work_package_wip_tokens WHERE id = 'wip_1'"
         )["state"] == "held"
-        with pytest.raises(sqlite3.IntegrityError, match="immutable"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="immutable"):
             store.execute(
                 "UPDATE work_package_lease_expiry_repairs SET reason = 'forged' "
                 "WHERE id = 'repair_lease_1'"
@@ -1052,7 +1052,7 @@ def test_terminal_expiry_repair_cancels_wip_before_node() -> None:
             "UPDATE tasks SET state = 'failed', owner_agent_id = NULL, lease_id = NULL, "
             "leased_until = NULL WHERE id = 'task_node'"
         )
-        with pytest.raises(sqlite3.IntegrityError, match="cancel exact held WIP"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="cancel exact held WIP"):
             store.execute(
                 "UPDATE work_package_task_links SET node_state = 'cancelled' "
                 "WHERE task_id = 'task_node'"
@@ -1086,7 +1086,7 @@ def test_batch_input_is_bound_to_exact_assignment_and_evidence_attempt() -> None
         _insert_package(store)
         _insert_link_and_assignment(store)
         _insert_evidence(store, "ev_attempt", "task_node")
-        with pytest.raises(sqlite3.IntegrityError, match="does not match"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="does not match"):
             _insert_attempt_link(
                 store,
                 evidence_id="ev_attempt",
@@ -1100,18 +1100,18 @@ def test_batch_input_is_bound_to_exact_assignment_and_evidence_attempt() -> None
             task_id="task_node",
             lease_id="lease_1",
         )
-        with pytest.raises(sqlite3.IntegrityError, match="immutable"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="immutable"):
             store.execute(
                 "UPDATE evidence_attempt_links SET attempt_ref = ? WHERE evidence_id = ?",
                 ("refs/mac/other", "ev_attempt"),
             )
-        with pytest.raises(sqlite3.IntegrityError, match="immutable"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="immutable"):
             store.execute(
                 "UPDATE evidence_attempt_verifications SET verifier = ? "
                 "WHERE evidence_id = ?",
                 ("forged", "ev_attempt"),
             )
-        with pytest.raises(sqlite3.IntegrityError, match="append-only"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="append-only"):
             store.execute(
                 "DELETE FROM evidence_attempt_verifications WHERE evidence_id = ?",
                 ("ev_attempt",),
@@ -1189,7 +1189,7 @@ def test_batch_input_is_bound_to_exact_assignment_and_evidence_attempt() -> None
             "ev_attempt",
             "now",
         )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             store.execute(batch_input_sql, batch_input_params)
         store.execute(
             "UPDATE work_package_node_candidates SET status = ?, accepted_at = ?, "
@@ -1236,17 +1236,17 @@ def test_batch_input_is_bound_to_exact_assignment_and_evidence_attempt() -> None
             "UPDATE work_package_integration_batches SET state = ? WHERE id = ?",
             ("assembling", "batch_inputs"),
         )
-        with pytest.raises(sqlite3.IntegrityError, match="state transition"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="state transition"):
             store.execute(
                 "UPDATE work_package_integration_batches SET state = ? WHERE id = ?",
                 ("queued", "batch_inputs"),
             )
-        with pytest.raises(sqlite3.IntegrityError, match="immutable"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="immutable"):
             store.execute(
                 "UPDATE work_package_batch_inputs SET ordinal = 2 WHERE id = ?",
                 ("input_1",),
             )
-        with pytest.raises(sqlite3.IntegrityError, match="immutable"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="immutable"):
             store.execute(
                 "DELETE FROM work_package_batch_inputs WHERE id = ?", ("input_1",)
             )
@@ -1287,12 +1287,12 @@ def test_integration_batch_fence_is_monotonic() -> None:
             "UPDATE work_package_integration_batches SET state = ? WHERE id = ?",
             ("assembling", "batch_fence"),
         )
-        with pytest.raises(sqlite3.IntegrityError, match="identity"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="identity"):
             store.execute(
                 "UPDATE work_package_integration_batches SET target_ref = ? WHERE id = ?",
                 ("refs/heads/other", "batch_fence"),
             )
-        with pytest.raises(sqlite3.IntegrityError, match="fence cannot decrease"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="fence cannot decrease"):
             store.execute(
                 "UPDATE work_package_integration_batches SET lease_fence = 3 WHERE id = ?",
                 ("batch_fence",),
@@ -1309,7 +1309,7 @@ def test_integration_batch_fence_is_monotonic() -> None:
             )["lease_fence"]
             == 4
         )
-        with pytest.raises(sqlite3.IntegrityError, match="new fence"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="new fence"):
             store.execute(
                 "UPDATE work_package_integration_batches "
                 "SET lease_owner = ?, lease_expires_at = ? WHERE id = ?",
@@ -1320,7 +1320,7 @@ def test_integration_batch_fence_is_monotonic() -> None:
             "SET lease_owner = ?, lease_expires_at = ?, lease_fence = ? WHERE id = ?",
             ("integrator_2", "later", 5, "batch_fence"),
         )
-        with pytest.raises(sqlite3.IntegrityError, match="current fence"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="current fence"):
             store.execute(
                 "UPDATE work_package_integration_batches SET candidate_sha = ?, "
                 "candidate_tree_digest = ?, candidate_ref = ?, candidate_fence = ? "
@@ -1345,7 +1345,7 @@ def test_integration_batch_fence_is_monotonic() -> None:
                 "batch_fence",
             ),
         )
-        with pytest.raises(sqlite3.IntegrityError, match="current fence"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="current fence"):
             store.execute(
                 "UPDATE work_package_integration_batches SET candidate_sha = ? "
                 "WHERE id = ?",
@@ -1386,7 +1386,7 @@ def test_node_candidate_requires_same_task_lease_and_evidence_attempt() -> None:
             task_id="task_other",
             lease_id="lease_other",
         )
-        with pytest.raises(sqlite3.IntegrityError):
+        with pytest.raises((StoreError, sqlite3.IntegrityError)):
             store.execute(
                 "INSERT INTO work_package_node_candidates ("
                 "id, task_id, package_id, plan_version, epoch, node_key, node_generation, "
@@ -1436,7 +1436,7 @@ def test_node_candidate_requires_same_task_lease_and_evidence_attempt() -> None:
             "UPDATE work_package_task_links SET node_state = ? WHERE task_id = ?",
             ("candidate_submitted", "task_node"),
         )
-        with pytest.raises(sqlite3.IntegrityError, match="metadata"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="metadata"):
             store.execute(
                 "UPDATE work_package_node_candidates SET status = ?, accepted_at = ? "
                 "WHERE id = ?",
@@ -1454,7 +1454,7 @@ def test_node_candidate_requires_same_task_lease_and_evidence_attempt() -> None:
         assert get_work_package_task_link(store, "task_node").node_state == (
             "candidate_accepted"
         )
-        with pytest.raises(sqlite3.IntegrityError, match="metadata"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="metadata"):
             store.execute(
                 "UPDATE work_package_node_candidates SET accepted_by = ? WHERE id = ?",
                 ("forged", "candidate_1"),
@@ -1505,7 +1505,7 @@ def test_candidate_acceptance_ignores_legacy_mutable_verification_claims() -> No
             "UPDATE work_package_task_links SET node_state = ? WHERE task_id = ?",
             ("candidate_submitted", "task_node"),
         )
-        with pytest.raises(sqlite3.IntegrityError, match="controller-verified"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="controller-verified"):
             store.execute(
                 "UPDATE work_package_node_candidates SET status = ?, accepted_at = ?, "
                 "accepted_by = ? WHERE id = ?",
@@ -1561,12 +1561,12 @@ def test_rejected_candidate_is_immutable_and_new_attempt_gets_new_candidate() ->
             "UPDATE work_package_task_links SET node_state = ? WHERE task_id = ?",
             ("rejected", "task_node"),
         )
-        with pytest.raises(sqlite3.IntegrityError, match="newer bounded assignment"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="newer bounded assignment"):
             store.execute(
                 "UPDATE work_package_task_links SET node_state = ? WHERE task_id = ?",
                 ("executing", "task_node"),
             )
-        with pytest.raises(sqlite3.IntegrityError, match="metadata"):
+        with pytest.raises((StoreError, sqlite3.IntegrityError), match="metadata"):
             store.execute(
                 "UPDATE work_package_node_candidates SET rejection_reason = ? WHERE id = ?",
                 ("forged", "candidate_attempt_1"),
